// web/platform-admin-web/src/api/generated/model/invitePageItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */
import type { InvitePageItemStatus } from './invitePageItemStatus';

export interface InvitePageItem {
  /** 초대 식별자 */
  id: string;
  /** 초대 대상 이름 */
  name: string;
  /** 초대 대상 이메일 */
  email: string;
  /** 초대 상태 */
  status: InvitePageItemStatus;
  /** 권한 식별자 */
  roleId: string;
  /** 권한 코드 */
  roleCode: string;
  /** 권한 이름 */
  roleName: string;
  /**
     * 초대 메모
     * @nullable
     */
  note: string | null;
  /** 생성 일시 */
  createdAt: string;
}