// web/platform-admin-web/src/api/generated/model/invitePageItemStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * 초대 상태
 */
export type InvitePageItemStatus = typeof InvitePageItemStatus[keyof typeof InvitePageItemStatus];


export const InvitePageItemStatus = {
  QUEUED: 'QUEUED',
  PENDING: 'PENDING',
  EXPIRED: 'EXPIRED',
  CANCELED: 'CANCELED',
  ACCEPTED: 'ACCEPTED',
  REJECTED: 'REJECTED',
} as const;