// web/platform-admin-web/src/api/generated/model/membersControllerGetInvitePageV1DirectionItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */

export type MembersControllerGetInvitePageV1DirectionItem = typeof MembersControllerGetInvitePageV1DirectionItem[keyof typeof MembersControllerGetInvitePageV1DirectionItem];


export const MembersControllerGetInvitePageV1DirectionItem = {
  asc: 'asc',
  desc: 'desc',
} as const;