// web/platform-admin-web/src/api/generated/model/membersControllerGetInvitePageV1SortItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */

export type MembersControllerGetInvitePageV1SortItem = typeof MembersControllerGetInvitePageV1SortItem[keyof typeof MembersControllerGetInvitePageV1SortItem];


export const MembersControllerGetInvitePageV1SortItem = {
  createdAt: 'createdAt',
} as const;